Add files via upload
Hello this did not turn out as well as I hoped as I got stuck several times and simply ran out of time to do everything I would've liked with it. I tried to animate the pendulums but could not figure that out and became frustrated and I hope this is at least enough to pass the final and the class. Thank you for this semester I have learned a great deal and it was an enjoyable experience and I just hope I have done enough here. Have a fantastic summer